+++ foureng/products/__init__.py
@@ -23,6 +23,7 @@ from .forward_start import ForwardStartOption
 from .lookback import LookbackOption
 from .multi_asset import BasketOption, BestOfOption, ExchangeOption, SpreadOption
 from .parisian import ParisianOption
+from .quanto import QuantoOption
 from .variance import VarianceOption, VarianceSwap
 
 __all__ = [
@@ -46,4 +47,5 @@ __all__ = [
     "ParisianOption",
     "CompoundOption",
     "ChooserOption",
+    "QuantoOption",
 ]